% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{censorDateDoc}
\alias{censorDateDoc}
\title{Helper for consistent documentation of \code{censorDate}.}
\arguments{
\item{censorDate}{Name of the date column in \code{cohort} that identifies the end
of follow-up, inclusive. If \code{NULL}, the end of each individual's observation
period is used.}
}
\description{
Helper for consistent documentation of \code{censorDate}.
}
\keyword{internal}
