% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cohortDoc}
\alias{cohortDoc}
\title{Helper for consistent documentation of \code{cohort}.}
\arguments{
\item{cohort}{A \code{cohort_table} object containing one cohort for each state in
the multi-state model. Cohort names must match the state names in \code{trans}.}
}
\description{
Helper for consistent documentation of \code{cohort}.
}
\keyword{internal}
