% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{eventDateDoc}
\alias{eventDateDoc}
\title{Helper for consistent documentation of \code{eventDate}.}
\arguments{
\item{eventDate}{Name of the date column in \code{cohort} that identifies when an
individual enters a state.}
}
\description{
Helper for consistent documentation of \code{eventDate}.
}
\keyword{internal}
