% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{followUpDaysDoc}
\alias{followUpDaysDoc}
\title{Helper for consistent documentation of \code{followUpDays}.}
\arguments{
\item{followUpDays}{Maximum number of days of follow-up for which to estimate
state occupation probabilities. Use \code{Inf} to include all available
follow-up.}
}
\description{
Helper for consistent documentation of \code{followUpDays}.
}
\keyword{internal}
