% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{keepExtraColumnsDoc}
\alias{keepExtraColumnsDoc}
\title{Helper for consistent documentation of \code{keepExtraColumns}.}
\arguments{
\item{keepExtraColumns}{Logical value indicating whether additional columns
from \code{cohort} should be retained in the prepared multi-state data.}
}
\description{
Helper for consistent documentation of \code{keepExtraColumns}.
}
\keyword{internal}
