% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopMultistate-package.R
\docType{package}
\name{OmopMultistate-package}
\alias{OmopMultistate}
\alias{OmopMultistate-package}
\title{OmopMultistate: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopMultistate/}
  \item Report bugs at \url{https://github.com/oxford-pharmacoepi/OmopMultistate/issues}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
