% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMultiStateProbabilities.R
\name{plotMultistateProbabilities}
\alias{plotMultistateProbabilities}
\title{Title}
\usage{
plotMultistateProbabilities(
  result,
  facet = "cdm_name",
  colour = "state",
  style = NULL,
  initialState = NULL
)
}
\arguments{
\item{initialState}{}
}
\description{
Title
}
