% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMultistateData.R
\name{prepareMultistateData}
\alias{prepareMultistateData}
\title{Title}
\usage{
prepareMultistateData(
  cohort,
  trans,
  eventDate = "cohort_start_date",
  censorDate = NULL,
  stateHierarchy = character(),
  stateStep = 0.01,
  keepExtraColumns = TRUE
)
}
\arguments{
\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. If S is the number of states in the multi-state model,
trans should be an S x S matrix, with (i,j)-element a positive integer if a
transition from i to j is possible in the multi-state model, NA otherwise. In
particular, all diagonal elements should be NA. The integers indicating the
possible transitions in the multi-state model should be sequentially
numbered, 1,...,K, with K the number of transitions.}

\item{keepExtraColumns}{}
}
\description{
Title
}
