% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{transMat}
\alias{tidy}
\alias{filterGroup}
\alias{filterStrata}
\alias{filterAdditional}
\alias{filterSettings}
\alias{filterResult}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mstate}{\code{\link[mstate]{transMat}}}

  \item{omopgenerics}{\code{\link[omopgenerics]{filterAdditional}}, \code{\link[omopgenerics]{filterGroup}}, \code{\link[omopgenerics]{filterResult}}, \code{\link[omopgenerics]{filterSettings}}, \code{\link[omopgenerics]{filterStrata}}, \code{\link[omopgenerics:reexports]{tidy}}}
}}

