% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{resultDoc}
\alias{resultDoc}
\title{Helper for consistent documentation of \code{result}.}
\arguments{
\item{result}{A \code{summarised_result} object produced by
\code{\link[=summariseMultistateProbabilities]{summariseMultistateProbabilities()}}.}
}
\description{
Helper for consistent documentation of \code{result}.
}
\keyword{internal}
