% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{stateHierarchyDoc}
\alias{stateHierarchyDoc}
\title{Helper for consistent documentation of \code{stateHierarchy}.}
\arguments{
\item{stateHierarchy}{Character vector of state names in the order used to
resolve states that occur on the same date. An empty character vector applies
no hierarchy.}
}
\description{
Helper for consistent documentation of \code{stateHierarchy}.
}
\keyword{internal}
