% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{stateStepDoc}
\alias{stateStepDoc}
\title{Helper for consistent documentation of \code{stateStep}.}
\arguments{
\item{stateStep}{Numeric increment used to separate states that occur on the
same date according to \code{stateHierarchy}.}
}
\description{
Helper for consistent documentation of \code{stateStep}.
}
\keyword{internal}
