% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{strataDoc}
\alias{strataDoc}
\title{Helper for consistent documentation of \code{strata}.}
\arguments{
\item{strata}{List of character vectors defining stratifications. Each
character vector identifies one or more columns in \code{cohort}; a separate model
is fitted for each stratification.}
}
\description{
Helper for consistent documentation of \code{strata}.
}
\keyword{internal}
