% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{styleDoc}
\alias{styleDoc}
\title{Helper for consistent documentation of \code{style}.}
\arguments{
\item{style}{Plot style passed to \code{\link[visOmopResults:themeVisOmop]{visOmopResults::themeVisOmop()}}. It can be
the name of a built-in style, a path to a YAML style file, or \code{NULL} to use
the default style.}
}
\description{
Helper for consistent documentation of \code{style}.
}
\keyword{internal}
