% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{timeScaleDoc}
\alias{timeScaleDoc}
\title{Helper for consistent documentation of \code{timeScale}.}
\arguments{
\item{timeScale}{Character string specifying the time scale of the x-axis:
either \code{"days"} or \code{"years"}.}
}
\description{
Helper for consistent documentation of \code{timeScale}.
}
\keyword{internal}
