% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{transDoc}
\alias{transDoc}
\title{Helper for consistent documentation of \code{trans}.}
\arguments{
\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. If S is the number of states in the multi-state model,
trans should be an S x S matrix, with (i,j)-element a positive integer if a
transition from i to j is possible in the multi-state model, NA otherwise. In
particular, all diagonal elements should be NA. The integers indicating the
possible transitions in the multi-state model should be sequentially
numbered, 1,...,K, with K the number of transitions.}
}
\description{
Helper for consistent documentation of \code{trans}.
}
\keyword{internal}
