% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionMatrix.R
\name{transitionMatrix}
\alias{transitionMatrix}
\title{Title}
\usage{
transitionMatrix(cohortNames = NULL, transitions = list())
}
\arguments{
\item{transitions}{}
}
\description{
Title
}
\examples{
library(OmopMultistate)

transitionMatrix(cohortNames = c("treated", "untreated", "dead"))

transitionMatrix(
  cohortNames = c("treated", "untreated", "dead"),
  transitions = list(
    c("treated", "untreated"), c("untreated", "treated"),
    c("treated", "dead"), c("untreated", "dead")
  )
)

transitionMatrix(
  cohortNames = c("treated", "untreated", "dead"),
  transitions = list(c(1, 2), c(2, 1), c(1, 3), c(2, 3))
)

transitionMatrix(
  transitions = list(
    c("treated", "untreated"), c("untreated", "treated"),
    c("treated", "dead"), c("untreated", "dead")
  )
)

}
