% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopMultistate-package.R
\docType{package}
\name{OmopMultistate-package}
\alias{OmopMultistate}
\alias{OmopMultistate-package}
\title{OmopMultistate: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopMultistate/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
