% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMultistateProbabilities.R
\name{plotMultistateProbabilities}
\alias{plotMultistateProbabilities}
\title{Title}
\usage{
plotMultistateProbabilities(result, style = NULL, timeScale = "days")
}
\arguments{
\item{timeScale}{}
}
\description{
Title
}
