% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMultiStateProbabilities.R
\name{plotMultistateProbabilities}
\alias{plotMultistateProbabilities}
\title{Plot Multi-State Occupation Probabilities extracted by
\code{summariseMultistateProbabilities()}}
\usage{
plotMultistateProbabilities(result, style = NULL, timeScale = "days")
}
\arguments{
\item{result}{A \code{summarised_result} object produced by
\code{\link[=summariseMultistateProbabilities]{summariseMultistateProbabilities()}}.}

\item{style}{Plot style passed to \code{\link[visOmopResults:themeVisOmop]{visOmopResults::themeVisOmop()}}. It can be
the name of a built-in style, a path to a YAML style file, or \code{NULL} to use
the default style.}

\item{timeScale}{Character string specifying the time scale of the x-axis:
either \code{"days"} or \code{"years"}.}
}
\value{
A ggplot2 object with the visualisation of the probabilities over
time.
}
\description{
Plot Multi-State Occupation Probabilities extracted by
\code{summariseMultistateProbabilities()}
}
