% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMsData.R
\name{prepareMultiStateData}
\alias{prepareMultiStateData}
\title{Title}
\usage{
prepareMultiStateData(
  cohort,
  tmat,
  eventDate = "cohort_start_date",
  censorDate = NULL,
  stateHierarchy = character(),
  stateStep = 0.01,
  keepExtraColumns = TRUE
)
}
\arguments{
\item{keepExtraColumns}{}
}
\description{
Title
}
