% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMultiStateProbabilities.R
\name{summariseMultistateProbabilities}
\alias{summariseMultistateProbabilities}
\title{Title}
\usage{
summariseMultistateProbabilities(
  cohort,
  trans,
  strata = list(),
  followUpDays = Inf,
  eventDate = "cohort_start_date",
  censorDate = NULL,
  stateHierarchy = character(),
  stateStep = 0.01
)
}
\arguments{
\item{strata}{A different model is fit for each stratification.}

\item{stateStep}{}
}
\description{
Title
}
