% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseMultistateProbabilities.R
\name{summariseMultistateProbabilities}
\alias{summariseMultistateProbabilities}
\title{Title}
\usage{
summariseMultistateProbabilities(
  cohort,
  trans,
  strata = list(),
  followUpDays = Inf,
  eventDate = "cohort_start_date",
  censorDate = NULL,
  stateHierarchy = character(),
  stateStep = 0.01
)
}
\arguments{
\item{cohort}{A \code{cohort_table} object containing one cohort for each state in
the multi-state model. Cohort names must match the state names in \code{trans}.}

\item{trans}{Transition matrix describing the states and transitions in the
multi-state model. If S is the number of states in the multi-state model,
trans should be an S x S matrix, with (i,j)-element a positive integer if a
transition from i to j is possible in the multi-state model, NA otherwise. In
particular, all diagonal elements should be NA. The integers indicating the
possible transitions in the multi-state model should be sequentially
numbered, 1,...,K, with K the number of transitions.}

\item{strata}{List of character vectors defining stratifications. Each
character vector identifies one or more columns in \code{cohort}; a separate model
is fitted for each stratification.}

\item{followUpDays}{Maximum number of days of follow-up for which to estimate
state occupation probabilities. Use \code{Inf} to include all available
follow-up.}

\item{eventDate}{Name of the date column in \code{cohort} that identifies when an
individual enters a state.}

\item{censorDate}{Name of the date column in \code{cohort} that identifies the end
of follow-up. If \code{NULL}, the end of each individual's observation period is
used.}

\item{stateHierarchy}{Character vector of state names in the order used to
resolve states that occur on the same date. An empty character vector applies
no hierarchy.}

\item{stateStep}{Numeric increment used to separate states that occur on the
same date according to \code{stateHierarchy}.}
}
\value{
A \verb{<summarised_result>} object with result_type =
"summarise_multistate_probabilities" that contains the multistate
probabilities.
}
\description{
Title
}
